requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
fastapi==0.112.0
h11==0.14.0
idna==3.7
importlib_metadata==8.4.0
json-logging==1.3.0
loki-logger-handler==0.1.4
opentelemetry-api==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
prometheus_client==0.20.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.1.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
wrapt==1.16.0
zipp==3.20.2
